more refactors

+++ linear-model_test.cc
@@ -19,13 +19,13 @@ TEST(LinearRegression, Basic) {
 
 TEST(LogisticRegression, Basic) {
   Eigen::MatrixXd X(4, 2);
-  Eigen::VectorXd y(4);
+  Eigen::VectorXi y(4);
   X << 1, 1, 1, 2, 1, 3, 1, 4;
   y << 0, 0, 1, 1;
 
   auto lr = LogisticRegression(0.1);
   lr.fit(X, y);
 
-  EXPECT_NEAR(0.0, mean_squared_error(lr.predict(X), y), 1e-5);
+  EXPECT_NEAR(0.0, mean_squared_error(lr.predict(X), y.cast<double>()), 1e-5);
 }
 }
